requirements.txt
sqlalchemy~=2.0.41
fastapi~=0.115.12
jose~=1.0.0
passlib~=1.7.4
pydantic~=2.11.5
uvicorn~=0.34.2
redis~=6.1.0
python-dotenv~=1.1.0